bottomSheetStateをファイルに分離
